more game data including shop & pin affinity elements

+++ ntwewy-db/Classes/SkillTree.cs
@@ -9,7 +9,7 @@
         public int InfoUpdateScenario { get; set; }
         public string DayText { get; set; }
         public string PlaceText { get; set; }
-        public int ShopId { get; set; }
+        public sbyte ShopId { get; set; }
         public int ReleaseValue { get; set; }
         public int Board { get; set; }
         public int SkillId { get; set; }
@@ -19,7 +19,7 @@
         public short ParentId { get; set; }
         public int SaveIndex { get; set; }
 
-        public SkillTree(short id, string characterName, string[] characterInfo, bool infoUpdateIfConnect, int infoUpdateScenario, string dayText, string placeText, int shopId, int releaseValue, int board, int skillId, string characterIcon, int entryDay, int connectDay, short parentId, int saveIndex)
+        public SkillTree(short id, string characterName, string[] characterInfo, bool infoUpdateIfConnect, int infoUpdateScenario, string dayText, string placeText, sbyte shopId, int releaseValue, int board, int skillId, string characterIcon, int entryDay, int connectDay, short parentId, int saveIndex)
         {
             Id = id;
             CharacterName = characterName;
